Type mismatch fixes

+++ Include/RenderingEngine/OpenGL/OpenGL.hpp
@@ -33,7 +33,7 @@ namespace RenderingEngine
             void BindFramebuffer(const OpenGL::Framebuffer& framebuffer);
             void BindFramebuffer();
 
-            void DrawArrays(const OpenGL::VertexArray& vao, OpenGL::Primitive mode, unsigned int offset, unsigned int vertices);
+            void DrawArrays(const OpenGL::VertexArray& vao, OpenGL::Primitive mode, unsigned int offset, size_t vertices);
             void DrawElements(const OpenGL::VertexArray& vao, OpenGL::Primitive mode, intptr_t offset, unsigned int count, unsigned int type);
         };
     }
